requirements.txt
streamlit
pandas
openpyxl
plotly
python-dotenv
cryptography